+++ server/camcops_server/cc_modules/cc_patient.py
@@ -1207,6 +1207,29 @@ class Patient(GenericTabletRecordMixin, Base):
 
         return proquint_from_uuid(self.uuid)
 
+    @property
+    def duplicates(self) -> set["Patient"]:
+        """
+        Returns a set of patients that have an ID Number that matches this one.
+        For a patient to be considered a duplicate, the ID number needs to
+        match on:
+
+        * Group
+        * Type of ID Number (e.g. NHS Number)
+        * Value of ID number
+        * Device where patient was created
+
+        and have the "current" flag set to True
+        """
+
+        dups = set()
+
+        for idnum in self.idnums:
+            for dup in idnum.duplicates:
+                dups.add(dup.patient)
+
+        return dups
+
 
 # =============================================================================
 # Validate candidate patient info for upload